paul bunyan is updated with images
Added all the different parts to paul bunyan,
Anchored all of the different components to have the correct
positioning. They are all anchored/centered correctly.

note: I changed the Paul bunyan variable from this.sprite inside paul.js
to this.body

I want the "body" to be the parent of all the different parts of him.