adding tweens for walking animation
so the anchor of paul is the this.core sprite that is created inside
paul.js

It is a 8x8 png file that is transparent, so everything is anchored
around thetransparent "core". This lets me tween everybody part.

So each body part is added to the "core" and their position is relative
to that.

You can tween:
Neck
Head
Body
Mustache
Left/Right Legs