fixing chopping area, and layering
trying to fix the layering of objects... I have it were the trees are
always drawn in the first array slot so Paul will be above them in the
layering

Also I changed the area for chopping because it was wrong. It was based
off his core but if he would stand at the top half of the circle it
would stop chopping.

Now I placed the get.bounds.contains(x,y) to be at his feet so it will
check against his feet instead of his core. It works at all spots inside
the circle now.