tree is broken
I made it were you can't get anymore lumber if you walk back into the
circle after its deleted.

The error is that the area bounds aren't being deleted or clear, so that
rectangle is still there even though the area sprite got deleted. So
this.chop() -> TrunkShake.start() both get called and then it wants to
delete objects again even though they are already deleted and it wanted
to add more lumber but I threw an if statement to stop the lumber adding

-added alpha channel to only have the circle appear if you are inside of
it.