added the wolf js, pngs
-added the wolf.js to the index

-added a test wolf to the playscreen

-i have the wolf loaded and animated with its walkingtweens. I had to
use WolfWalkingTweens because game.walkingtweens is gobal so i had to
make the wolf its own.

- the wolf tween animation is pretty good except the back legs kinda
look funny, I may tweak that a bit later

-Also I used the same concept as paul for the tween animations.
I created a core for the wolf which is the parent of all the body
parents. So if you want to move the wolf use the core.position to move
the entire wolf