"Fix" end of game crash
Put fix in quotes because this is a hack.  Basically when we transition
from play screen to win screen we remove all sprites from the stage.
One of these sprites is a piece of paper which has a tween.  The tween
keeps ticking and onComplete it tries to remove itself again, but it
has been previously removed by the resetStage.

To work around this for now I am only removing the sprite in the tween
if it exists in the children array on the stage.