bears deal 10 damage,  end game bug fixed
-The bears attack every 2 seconds, deal 10 damage. Changed the spawn
rate back to 8 seconds, i had it at 1 second for testing/debugging
reasons
-Also I think I fixed the bug when you kill an animal and win the game
at the same time. It won't break and try to find the index of the animal
inside the death tween anymore

I have it when an enemy dies, the enemy boolean goes to false then the
update loop checks if an animal has enemy = false, if so remove it from
the enemies array.

That seemed to fix the bug, I cannot recreate it again