added controls,info
-added the information on how to win the game
'3 Lumber = 1 Paper     100 Paper to Win'

-added the controls 'Move: WASD   Fire : Leftclick'

I think that is good enough for the explanation? if not change it up or
add more

I think people will realize they stand in the trees to get lumber then
bring to the mill

- edited the win condition for it needs 100 paper to win. Each tree
gives 24 lumber, 24/3 = 8. So each tree gives 8 lumber. 100/8 = 12.5

so you need to harvest 12.5 trees into paper to win the game